merge branch lineage-22.2 into aospb-Q2 (#6)
* Change TARGET_RECOVERY_UI_LIB to use fully qualified names in coral

Since the TARGET_RECOVERY_UI_LIB for each device is not exactly the
same, and most of the modules in it contain `soong_namespace{}` to
avoid conflicts between different devices, after converting
`librecovery_ui_ext` from Android.mk to Android.bp, it needs to be
changed to a `fully qualified name` in order to correctly use the
corresponding module.

Flag: EXEMPT refactor

Bug: 339143524
Test: m -j librecovery_ui_ext
Change-Id: I3b727d389b1ee2c8df87010e3aedb04175dd4e93

* [owners] Remove migueldedios@google.com from OWNERS

This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://android-review.googlesource.com/q/migueldedios@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ie3436522b1b68797abfed7e958c72e361ad1be42

* [owners] Remove aclin@google.com from OWNERS

This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://android-review.googlesource.com/q/aclin@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I93a6b2af365f166eeb8b5ba0923b1d747c87f120

* [owners] Remove etancohen@google.com from rro_overlays/WifiOverlay/OWNERS

This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://android-review.googlesource.com/q/etancohen@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I84f0b94a01ab4318ad9bef9a29488feecbe3bed9

* floral: switch to common QCOM AIDL bootctrl HAL

Change-Id: I04ab771d3b1c38b58913607fbff1bb3b55e1fe25

* floral: Don't use BSG framework for gpt-utils

Change-Id: Ia6e5d651560c8ae5186a48bfbb0e30b864150564

* floral: sepolicy: Add default permission for aidl hal_bootctl

add in default sepolicy configuration for aidl implementation
of hal_bootctl

Change-Id: I8bd86071dde93d8ed5ea41555b6d3ede9d434838

* floral: Enable ELF checks for libwpa_client dependent libs

Change-Id: I490df02b68da4b9a7088a239672a1dcace66d270

* floral: Shim color_init with libbase_shim

See: https://cs.android.com/android/_/android/platform/system/libbase/+/46fc805282265715266c256fbe34cf1a886c475f

Change-Id: I5127020e18114a13515d9dc382c95d689634243d

* floral: Import cs40l25 vibrator HAL from AP4A

Change-Id: Ia58db0eefb34ac47081b0f1483bad64687dd295e

* floral: Set chre_daemon_dsp_library

Change-Id: If0a2be3eee2f55fcbb48ca9118f4f14f68ac1c32

* floral: Update CHRE daemon name

Change-Id: Ia1139714c5019684b1e62933fd34632aca2e02a7

* fixup! Change TARGET_RECOVERY_UI_LIB to use fully qualified names in coral

Change-Id: Ie0e77bd84e757593124a71ff34422ea682f595fb

* coral: Remove vendor RenderScript implementation.

 * RenderScript is deprecated on newer platforms and
   is being officially replaced.

 * On April 19, 2021, Google announced that RenderScript
   will be deprecated in Android 12, and recommended
   porting existing code to Vulkan.

   https://android-developers.googleblog.com/2021/04/android-gpu-compute-going-forward.html

Change-Id: I19460ef266a646b046f1e7d2f0b4eab7c48ae536

* floral: Remove CleanSpec.mk

Change-Id: I6af355cfbfd5fc388ca25cd25cea9df9e74c41c3

* floral: Migrate gralloc_handle_has_reserved_size to soong_config_set

Change-Id: I3aa081f081e41ba108ff91227409baad7f54d49c

* floral: Migrate Lineage Health to soong_config_set

Change-Id: I2eff39e83b9673aa2989d05de65c646241ba61da

---------

Co-authored-by: Nelson Li <nelsonli@google.com>
Co-authored-by: Owner Cleanup Bot <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Co-authored-by: Cosmin Tanislav <demonsingur@gmail.com>
Co-authored-by: Michael Bestas <mkbestas@lineageos.org>
Co-authored-by: Anthony Adamo <quic_aadamo@quicinc.com>
Co-authored-by: Michael Bestas <mkbestas@gmail.com>
Co-authored-by: Alexander Koskovich <zvnexus@outlook.com>
Co-authored-by: Bruno Martins <bgcngm@gmail.com>